Add a --exact option
Examine the disassembly for padding bytes and subtract those from the
function size.  This provides a more accurate size difference as
otherwise we can only tell differences when a function's size crosses
the alignment (e.g. removed a few bytes of code, but symbol size stays
the same due to 8 byte alignment of functions)